chore(orga): moved every method in parent component
